+++ src/components/chat/ChatContents.tsx
@@ -9,17 +9,17 @@ type Props = {
 };
 const ChatContents = ({ chatContentList }: Props) => {
   const myId = 'myNickname'; // 전역변수에서 가져온다.
+
+  const chatItem = (chatContent: ChatInterface) => {
+    if (chatContent.nickname === myId) {
+      return <MyChat chatContent={chatContent} />;
+    } else {
+      return <YourChat chatContent={chatContent} />;
+    }
+  };
   return (
     <ChatContentsLayout>
-      {chatContentList.map((chatContent) => (
-        <ChatItemBox key={chatContent.chatId}>
-          {chatContent.nickname === myId ? (
-            <MyChat chatContent={chatContent} />
-          ) : (
-            <YourChat chatContent={chatContent} />
-          )}
-        </ChatItemBox>
-      ))}
+      {chatContentList.map((chatContent) => chatItem(chatContent))}
     </ChatContentsLayout>
   );
 };
@@ -29,10 +29,7 @@ const ChatContentsLayout = styled.div`
   width: 100%;
   gap: 0.8rem;
   margin: 0 0 4rem 0;
-
   display: flex;
   flex-direction: column;
   overflow-y: scroll;
 `;
-
-const ChatItemBox = styled.div``;
