+++ src/components/search/BoldContent.tsx
